Relax ThenFunc signature to allow any HandlerFunc
Relax ThenFunc signature to allow any HandlerFunc (based on the
anonymous func type of HandlerFunc) not just those with the
http.HandlerFunc concrete type.
This gives more flexibility to users of the chain API.

While this can be considered a breaking change (the ThenFunc signature
is not strictly the same), the new signature widens the accepted input
values for ThenFunc.